[COOK-2043] Default to "git" on recent Debianoids
Debian 6.0 "squeeze" and Ubuntu 10.10 "maverick" renamed the package to
"git" and git-core is a transitional dummy package. On some point it
will be removed.

Only install "git-core" on ancient Ubuntu versions as Debian 5.0 hasn't
been supported for a while. Ubuntu 10.04 LTS still is.

This commit is based on a commit by Deepak Kannan <kannan.deepak@gmail.com>.