Fix `package_id` for Mac platforms to prevent perpetual reinstalls.
This commit updates the `default['git']['osx_dmg']['package_id']` value
which is currently incorrect. The result is that the dmg resource will
always fail its `installed?` check and re-install this packge on each
Chef run.

The correct package_id was determined by running:

    > pkgutil --pkgs | grep GitOSX
    GitOSX.Installer.git195Universal.etc.pkg
    GitOSX.Installer.git195Universal.git.pkg

The result is a substantial speed improvement when running this cookbook
on Mac platforms.